add refact_utils module and move scripts into it
