+++ refact_webgui/webgui/selfhost_login.py
@@ -10,7 +10,7 @@ from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from pydantic import Required
 
-from self_hosting_machinery import env  # REFACTORME
+from refact_utils.scripts import env
 from refact_webgui.webgui import static_folders
 
 from typing import List
