+++ refact_webgui/webgui/selfhost_lsp_proxy.py
@@ -9,7 +9,7 @@ from starlette.requests import Request
 from starlette.responses import StreamingResponse
 from starlette.background import BackgroundTask
 
-from self_hosting_machinery import env  # REFACTORME
+from refact_utils.scripts import env
 from refact_webgui.webgui.selfhost_login import RefactSession
 
 __all__ = ["LspProxy"]
