+++ refact_webgui/webgui/tab_settings.py
@@ -5,7 +5,8 @@ from fastapi import APIRouter
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from pathlib import Path
-from self_hosting_machinery import env  # REFACTORME
+
+from refact_utils.scripts import env
 from refact_webgui.webgui.selfhost_model_assigner import ModelAssigner
 
 from typing import Optional
