+++ self_hosting_machinery/finetune/configuration/finetune_config.py
@@ -1,8 +1,8 @@
 import math
 import torch
 
+from refact_utils.scripts import env
 from self_hosting_machinery.finetune.utils import traces
-from self_hosting_machinery import env
 
 from typing import Any, Dict, List
