+++ self_hosting_machinery/finetune/scripts/aux/file_status_context.py
@@ -2,10 +2,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional, Any, Dict, List
 
+from refact_utils.scripts import env
 from self_hosting_machinery.finetune.scripts.aux.finetune_filter_status_tracker import FinetuneFilterStatusTracker
 from self_hosting_machinery.finetune.utils import traces
 from self_hosting_machinery.finetune.utils.finetune_utils import get_file_digest
-from self_hosting_machinery.scripts import env
 
 __all__ = ['FilesStatusContext', 'FileStatus']
