+++ self_hosting_machinery/finetune/scripts/aux/finetune_filter_status_tracker.py
@@ -3,9 +3,9 @@ import os
 import time
 from typing import Dict, Any, Optional
 
+from refact_utils.scripts import env
 from self_hosting_machinery.finetune.utils.eta import EtaTracker
 from self_hosting_machinery.finetune.utils.finetune_utils import get_finetune_filter_stat
-from self_hosting_machinery.scripts import env
 
 __all__ = ['FinetuneFilterStatusTracker']
