+++ self_hosting_machinery/finetune/scripts/finetune_train.py
@@ -10,7 +10,7 @@ from typing import Dict, Any, Iterable, Tuple
 
 import torch as th
 
-from self_hosting_machinery import env
+from refact_utils.scripts import env
 from self_hosting_machinery.finetune.configuration.finetune_config import base_config, ConfigBuilder
 from self_hosting_machinery.finetune.scripts.aux.dataset import (
     create_train_dataloader, create_test_dataloader, get_ds_len_per_epoch, to_cuda
